convert ExternalIssueNo column to numeric and order desc by it https://github.com/metasfresh/metasfresh/issues/5816

+++ backend/de.metas.serviceprovider/src/test/java/de/metas/serviceprovider/issue/IssueRepositoryTest.java
@@ -77,7 +77,7 @@ public class IssueRepositoryTest
 			.type(IssueType.EXTERNAL)
 			.isEffortIssue(true)
 			.processed(true)
-			.externalIssueNo(MOCK_EXTERNAL_ISSUE_NO)
+			.externalIssueNo(BigDecimal.valueOf(MOCK_EXTERNAL_ISSUE_NO))
 			.externalIssueURL(MOCK_EXTERNAL_URL)
 			.issueLabels(MOCK_LABELS)
 			.build();
